+++ changelog.md
@@ -2,6 +2,7 @@
 
 * `getEnv` will return an empty object instead of throwing when a tree was initialized without environment
 * exposed `typecheck(type, value)` as public api (will ignore environment flags)
+* Introduced `createActionTrackingMiddleware` to simplify the creation of middleware that support complex async processes
 
 # 1.0.0